Formatting and fixed fix_encoding
* Move fix_encoding to helpers.py so its doctest can be run without
having Zope2 available.
* Correct the formatting of the doctest so it does not violate PEP8
* Fix the data being compared in the doctest (there was a property of
type string with value 20)
* Assert that fix_encoding is only called in Python 2. It does not work
yet with Python 3
* Change the doctest of fix_encoding so it does not fail in Python 3,
even though the function does not work there yet
